# Vectorize float array multiplication