@EgorBo 2023-09-11
In this case it's indeed probably better to just use `Vector<>` instead of code duplication.