@neon-sunset 2023-09-11
(alternatively, you could just keep Vector256 unconditionally for .NET 8 which can unroll this into perfect 128x2 codegen, because ARM CPUs compensate lack of 256b-wide vectors with more SIMD units, which works surprisingly well with .NET V256 fallbacks for basic math operations (i.e. the ones that do element-wise logic and not masks or horizontal ops))