@KeterSCP 2023-09-13
So here are the results for `Vector<T>` implementation. Speedup is the same as with Vector128/Vector256, but without code duplication and with 128 bytes less generated code:

<details>

<summary>
Benchmark results
</summary>

```

BenchmarkDotNet v0.13.8, Windows 8 (6.2.9200.2215)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 8.0.100-preview.7.23376.3
  [Host]     : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT AVX2
  DefaultJob : .NET 7.0.10 (7.0.1023.36312), X64 RyuJIT AVX2


```
| Method             | ArraySize | Mean         | Error     | StdDev    | Ratio | Code Size |
|------------------- |---------- |-------------:|----------:|----------:|------:|----------:|
| **ScalarMultiply**     | **1**         |     **1.380 ns** | **0.0071 ns** | **0.0060 ns** |  **1.00** |      **57 B** |
| VectorizedMultiply | 1         |     1.496 ns | 0.0069 ns | 0.0064 ns |  1.08 |     283 B |
|                    |           |              |           |           |       |           |
| **ScalarMultiply**     | **10**        |     **2.697 ns** | **0.0215 ns** | **0.0201 ns** |  **1.00** |      **57 B** |
| VectorizedMultiply | 10        |     2.400 ns | 0.0064 ns | 0.0057 ns |  0.89 |     283 B |
|                    |           |              |           |           |       |           |
| **ScalarMultiply**     | **16**        |     **4.808 ns** | **0.0172 ns** | **0.0152 ns** |  **1.00** |      **57 B** |
| VectorizedMultiply | 16        |     2.294 ns | 0.0118 ns | 0.0098 ns |  0.48 |     283 B |
|                    |           |              |           |           |       |           |
| **ScalarMultiply**     | **1000**      |   **255.150 ns** | **0.2478 ns** | **0.2070 ns** |  **1.00** |      **57 B** |
| VectorizedMultiply | 1000      |    34.298 ns | 0.0993 ns | 0.0880 ns |  0.13 |     283 B |
|                    |           |              |           |           |       |           |
| **ScalarMultiply**     | **10000**     | **2,537.047 ns** | **9.5801 ns** | **8.9613 ns** |  **1.00** |      **57 B** |
| VectorizedMultiply | 10000     |   457.372 ns | 0.9506 ns | 0.8427 ns |  0.18 |     283 B |


</details>